Move Power Platform mocks to their own file
